backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-filter>=23.5
djangorestframework-simplejwt>=5.3.1
python-dotenv>=1.0.0
Pillow>=9.0.0
